Move images, added "scopes" instructions for data marketplace.
All of the images in this repository were moved to a centralized "images" folder in the base directory and all references in the README files were updated accordingly.  Also added instructions for update the scopes for the data marketplace search to include the "Starburst Bussiness Domain" and "Starburst Data Product" asset types.  These instructions were added to both the "quick instructions" on the README as well as the user guide PDF.